services/users/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
cassandra-driver==3.29.0
python-jose==3.3.0
httpx==0.26.0
pydantic==2.5.0
python-multipart==0.0.6
PyJWT==2.8.0
cryptography==42.0.0